Fix import/packages bugs and refactor code.
Split TimedOutTestsListener.buildThreadsDump
